mocks: move runnables/mocks to internal/mocks
The mocks package was sitting at a non-internal/ path, which made it
de-facto public surface even though it's never been advertised in
README, CHANGELOG, or examples and only test code in this repo imports
it. An external consumer could import it and lock the API.

Move the package under internal/ so the import path itself signals
"this is not part of the public surface" and Go tooling blocks
out-of-module importers.

Mechanical change:
- git mv runnables/mocks/{mocks,mocks_test}.go internal/mocks/
- update the docstring example inside mocks.go and the import inside
  mocks_test.go
- update import paths in 17 importing test files across supervisor/,
  runnables/composite/, runnables/httpcluster/

Package name stays as `mocks`, so every existing call site
(`mocks.NewMockRunnable(...)` etc.) keeps compiling unchanged.

Breaking change for any external consumer importing
github.com/robbyt/go-supervisor/runnables/mocks. None known. Justified
by pre-1.0 SemVer (latest tag v0.0.21).
