+++ runnables/composite/config_test.go
@@ -3,7 +3,7 @@ package composite
 import (
 	"testing"
 
-	"github.com/robbyt/go-supervisor/runnables/mocks"
+	"github.com/robbyt/go-supervisor/internal/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
