+++ runnables/composite/runner_run_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/robbyt/go-supervisor/internal/finitestate"
-	"github.com/robbyt/go-supervisor/runnables/mocks"
+	"github.com/robbyt/go-supervisor/internal/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
