+++ supervisor/shutdown_test.go
@@ -6,7 +6,7 @@ import (
 	"testing/synctest"
 	"time"
 
-	"github.com/robbyt/go-supervisor/runnables/mocks"
+	"github.com/robbyt/go-supervisor/internal/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
