+++ supervisor/state_test.go
@@ -3,7 +3,7 @@ package supervisor
 import (
 	"testing"
 
-	"github.com/robbyt/go-supervisor/runnables/mocks"
+	"github.com/robbyt/go-supervisor/internal/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
